pyproject.toml
[build-system]
requires = ["setuptools>=80.9.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fastfusion"
version = "0.0.1"
description = "FastFusion"
authors = [
    {name = "Michael Gilbert", email = "gilbertm@mit.edu"},
    {name = "Tanner Andrulis", email = "andrulis@mit.edu"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy>=2.2.0",
    "pandas>=2.2.0",
    "scipy>=1.15.0",
    "tqdm>=4.67.0",
    "pydantic>=2.0.0",
    "pydantic_core>=2.33.0",
    "ruamel.yaml>=0.18.0",
    "jinja2>=3.1.0",
    "islpy-barvinok == 2025.2.5",
    "sympy>=1.14.0",
    "paretoset>=1.2.5",
    "matplotlib>=3.10.0",
    "plotly>=6.1.0",
    "pydot>=4.0.0",
    "platformdirs>=4.3.0",
    "joblib>=1.5.1",
    "hwcomponents",
    "hwcomponents-adc",
    "hwcomponents-cacti",
    "hwcomponents-library",
    "hwcomponents-neurosim",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "flake8",
    "mypy",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
"*" = ["*.txt", "*.md", "*.yaml", "*.yml"]

[project.scripts]
# Add any console scripts here if needed